requirements.txt
# Web server
fastapi==0.115.0
uvicorn==0.30.6

# LLM + Agent framework
langchain==0.3.25
langgraph==0.4.1
langchain-anthropic==0.3.15
langchain-openai==0.3.12

# Vector DB
langchain-pinecone==0.2.7
pinecone-client==5.0.1

# Firebase
google-cloud-firestore==2.19.0
firebase-admin==6.6.0

# Observability
langsmith==0.3.42

# Utilities
python-dotenv==1.0.1
pydantic==2.10.0
